+++ hs-bindgen/CHANGELOG.md
@@ -48,8 +48,7 @@
   after the global variable. Extern anonymous declarations
   (e.g., `extern struct { .. } config;`) are rejected as unusable.
   * Generate an `IsArray` instance for each newtype of a type with an `IsArray`
-* Support unnamed bit-field declarations (of width zero), used to instruct the
-  compiler to not pack any more fields into the current storage unit.
+* Support unnamed bit-field declarations, used for padding.
 * Generate bindings for nested struct and union declarations even if we failed
   to generate bindings for the enclosing struct or union. See [PR
   #1849][pr-1849].